--- src/main/java/ch/njol/skript/agent/events/start/DelayStartEvent.java
@@ -1,46 +1,0 @@
-/*
- *   This file is part of Skript.
- *
- *  Skript is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
- *
- *  Skript is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
- *
- *  You should have received a copy of the GNU General Public License
- *  along with Skript. If not, see <https://www.gnu.org/licenses/>.
- *
- *
- * Copyright 2011-2019 Peter Güttinger and contributors
- *
- */
-
-package ch.njol.skript.agent.events.start;
-
-import ch.njol.skript.agent.AgentEvent;
-import ch.njol.skript.util.Timespan;
-
-/**
- * Occurs when a delay event is queued.
- *
- * @since 2.2-V13
- */
-public class DelayStartEvent extends AgentEvent {
-
-    /**
-     * the duration of the delay
-     */
-    public final Timespan duration;
-
-    /**
-     * only for internal use
-     */
-    public DelayStartEvent(final Timespan duration) {
-        this.duration = duration;
-    }
-
-}
