+++ src/main/java/ch/njol/skript/agents/defaults/TaskTrackerAgent.java
@@ -1,31 +1,32 @@
 /*
- *   This file is part of Skript.
  *
- *  Skript is free software: you can redistribute it and/or modify
- *  it under the terms of the GNU General Public License as published by
- *  the Free Software Foundation, either version 3 of the License, or
- *  (at your option) any later version.
+ *     This file is part of Skript.
  *
- *  Skript is distributed in the hope that it will be useful,
- *  but WITHOUT ANY WARRANTY; without even the implied warranty of
- *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *  GNU General Public License for more details.
+ *    Skript is free software: you can redistribute it and/or modify
+ *    it under the terms of the GNU General Public License as published by
+ *    the Free Software Foundation, either version 3 of the License, or
+ *    (at your option) any later version.
  *
- *  You should have received a copy of the GNU General Public License
- *  along with Skript. If not, see <https://www.gnu.org/licenses/>.
+ *    Skript is distributed in the hope that it will be useful,
+ *    but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *    GNU General Public License for more details.
+ *
+ *    You should have received a copy of the GNU General Public License
+ *    along with Skript. If not, see <https://www.gnu.org/licenses/>.
  *
  *
- * Copyright 2011-2019 Peter Güttinger and contributors
+ *   Copyright 2011-2019 Peter Güttinger and contributors
  *
  */
 
-package ch.njol.skript.agent.defaults;
+package ch.njol.skript.agents.defaults;
 
 import ch.njol.skript.Skript;
-import ch.njol.skript.agent.SkriptAgent;
-import ch.njol.skript.agent.TrackerAgent;
-import ch.njol.skript.agent.events.end.DelayEndEvent;
-import ch.njol.skript.agent.events.start.DelayStartEvent;
+import ch.njol.skript.agents.SkriptAgent;
+import ch.njol.skript.agents.TrackerAgent;
+import ch.njol.skript.agents.events.end.DelayEndEvent;
+import ch.njol.skript.agents.events.start.DelayStartEvent;
 import org.bukkit.command.CommandSender;
 
 import java.util.concurrent.TimeUnit;
