+++ src/main/java/ch/njol/skript/effects/Delay.java
@@ -25,9 +25,9 @@ package ch.njol.skript.effects;
 import ch.njol.skript.ScriptLoader;
 import ch.njol.skript.Skript;
 import ch.njol.skript.SkriptConfig;
-import ch.njol.skript.agent.SkriptAgent;
-import ch.njol.skript.agent.events.end.DelayEndEvent;
-import ch.njol.skript.agent.events.start.DelayStartEvent;
+import ch.njol.skript.agents.SkriptAgent;
+import ch.njol.skript.agents.events.end.DelayEndEvent;
+import ch.njol.skript.agents.events.start.DelayStartEvent;
 import ch.njol.skript.doc.Description;
 import ch.njol.skript.doc.Examples;
 import ch.njol.skript.doc.Name;
